requirements.txt
streamlit
langchain
transformers
datasets
torch
sentencepiece
python-docx
langchain-community